Rename `testgen` to `dev_scripts`
We rename the folder for clarity ("testgen" might also mean "generation
of test data"). We also rename the module to simply
`test_codegen` for the very same reason.

Additionally, the file `requirement.txt` is inlined in `setup.py` so
that the folder structure looks much more neat.
