+++ dev_scripts/setup.py
@@ -10,15 +10,10 @@ from setuptools import setup, find_packages
 
 # pylint: disable=redefined-builtin
 
-here = os.path.abspath(os.path.dirname(__file__))
-
-with open(os.path.join(here, "requirements.txt"), encoding="utf-8") as fid:
-    install_requires = [line for line in fid.read().splitlines() if line.strip()]
-
 setup(
-    name="aas-core3.0-csharp-testgen",
+    name="aas-core3.0-csharp-dev-scripts",
     version="0.0.1",
-    description="Generate the code of the unit tests on an AAS meta-model.",
+    description="Provide scripts for development of aas-core3.0-csharp.",
     url="https://github.com/aas-core-works/aas-core3.0-csharp",
     author="Marko Ristin",
     author_email="marko@ristin.ch",
@@ -31,9 +26,11 @@ setup(
     ],
     license="License :: OSI Approved :: MIT License",
     keywords="asset administration shell code generation industry 4.0 industrie i4.0",
-    packages=find_packages(exclude=["tests", "continuous_integration", "dev_scripts"]),
-    install_requires=install_requires,
-    py_modules=["aas_core_3_0_rc2_csharp_testgen"],
-    package_data={"aas_core_codegen": ["py.typed"]},
-    data_files=[(".", ["requirements.txt"])],
+    packages=find_packages(exclude=["tests", "continuous_integration"]),
+    install_requires=[
+        "aas-core-meta@git+https://github.com/aas-core-works/aas-core-meta@9ccd31e#egg=aas-core-meta",
+        "aas-core-codegen@git+https://github.com/aas-core-works/aas-core-codegen@fcc252c#egg=aas-core-codegen",
+        "aas-core3.0@git+https://github.com/aas-core-works/aas-core3.0-python@e232aa2#egg=aas-core3.0",
+    ],
+    py_modules=["test_codegen"],
 )
