+++ dev_scripts/test_codegen/generate_test_for_enhancing.py
@@ -27,7 +27,7 @@ from aas_core_codegen.csharp.common import (
     INDENT5 as IIIII,
 )
 
-from aas_core_3_0_csharp_testgen.common import load_symbol_table
+from test_codegen.common import load_symbol_table
 
 
 def main() -> int:
