+++ dev_scripts/test_codegen/generate_test_verification_of_enums.py
@@ -16,7 +16,7 @@ import aas_core_codegen.run
 from aas_core_codegen import intermediate
 from aas_core_codegen.common import Stripped
 
-from aas_core_3_0_csharp_testgen.common import load_symbol_table
+from test_codegen.common import load_symbol_table
 
 
 def main() -> int:
